// frontend/src/api/schema.d.ts
export interface components {
    schemas: {
// ...
        FlowDocument: {
            /** Text */
            text: string;
            /** Blocks */
            blocks: components["schemas"]["FlowBlock"][];
        };
// ...
        PageDocument: {
            /** Pages */
            pages: components["schemas"]["DocumentPage"][];
        };
// ...
        PreviousReview: {
            /** Results */
            results: components["schemas"]["SuggestionDTO"][];
            /** Userfeedback */
            userFeedback?: string | null;
        };
// ...
        ReviewRecord: {
            /** Runid */
            runId: string;
            /**
             * Createdat
             * Format: date-time
             */
            createdAt: string;
            /** Document */
            document: components["schemas"]["PageDocument"] | components["schemas"]["FlowDocument"];
            /** Materials */
            materials: components["schemas"]["ReferenceMaterial"][];
            previousReview?: components["schemas"]["PreviousReview"] | null;
            moduleResults: components["schemas"]["CompletedModules"];
            response: components["schemas"]["ReviewResponse"];
        };
        /** ReviewResponse */
        ReviewResponse: {
            status: components["schemas"]["ReviewStatus"];
            /** Errors */
            errors: components["schemas"]["ModuleErrorDTO"][];
            /** Materialreviews */
            materialReviews: components["schemas"]["MaterialReviewDTO"][];
            /** Results */
            results: components["schemas"]["SuggestionDTO"][];
        };
// ...
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}